feat(ai-tools): extract knowledge lookup core
